requirements.txt
agentscope
requests
python-dotenv
google-search-results
